requirements.txt
# requirements.txt
pytest==8.3.2
requests==2.32.3
pandas==2.2.2
plotly==5.22.0
pyyaml==6.0.1
openpyxl==3.1.2